Refactor dev db scripts to reduce chances of accidental secret leak. Make reset mean the same thing as rails.
