+++ package.json
@@ -7,6 +7,7 @@
     "start": "next start -- -p 10001",
     "dev": "next dev -p 10000",
     "db:reset": "ts-node src/db/scripts/reset.ts",
+    "db:recreate": "ts-node src/db/scripts/recreate.ts",
     "db:migrate": "ts-node src/db/scripts/migrate.ts",
     "db:seed": "ts-node src/db/scripts/seed.ts",
     "db:sql": "docker exec -it hurado-postgres psql -U hurado",
