+++ src/db/scripts/migrate.ts
@@ -1,34 +1,3 @@
-import * as path from "path";
-import { promises as fs } from "fs";
-import { Migrator, FileMigrationProvider, Kysely } from "kysely";
-import { SCRIPTS_ONLY_DATABASE } from "./do-not-import-this";
+import { __DO_NOT_IMPORT__DeveloperDatabaseScripts } from "./do-not-import-this";
 
-migrate_database(SCRIPTS_ONLY_DATABASE);
-
-async function migrate_database(db: Kysely<any>) {
-  const migrator = new Migrator({
-    db,
-    provider: new FileMigrationProvider({
-      fs,
-      path,
-      // This needs to be an absolute path.
-      migrationFolder: path.join(__dirname, "../migrations"),
-    }),
-  });
-
-  const { error, results } = await migrator.migrateToLatest();
-
-  results?.forEach((it) => {
-    if (it.status === "Success") {
-      console.info(`migration "${it.migrationName}" was executed successfully`);
-    } else if (it.status === "Error") {
-      console.error(`failed to execute migration "${it.migrationName}"`);
-    }
-  });
-
-  if (error) {
-    console.error("failed to migrate");
-    console.error(error);
-    process.exit(1);
-  }
-}
+__DO_NOT_IMPORT__DeveloperDatabaseScripts.migrate_database();
