+++ src/db/scripts/recreate.ts
@@ -1,3 +1,3 @@
 import { __DO_NOT_IMPORT__DeveloperDatabaseScripts } from "./do-not-import-this";
 
-__DO_NOT_IMPORT__DeveloperDatabaseScripts.reset_database();
+__DO_NOT_IMPORT__DeveloperDatabaseScripts.recreate_database();
